ctype-big5.c: ctype-cp932.c: ctype-gbk.c: ctype-mb.c: ctype-simple.c: ctype-sjis.c: ctype-ucs2.c: ctype-ujis.c: ctype-utf8.c:
  Adding explicit cast to return type
  in pointer substructions to avoid
  warnings from some compilers.


strings/ctype-big5.c:
  Adding explicit cast to return type
  in pointer substructions to avoid
  warnings from some compilers.
  Adding explicit cast to return type
  in pointer substructions to avoid
  warnings from some compilers.
  Adding explicit cast to return type
  in pointer substructions to avoid
  warnings from some compilers.
strings/ctype-cp932.c:
  Adding explicit cast to return type
  in pointer substructions to avoid
  warnings from some compilers.
strings/ctype-gbk.c:
  Adding explicit cast to return type
  in pointer substructions to avoid
  warnings from some compilers.
strings/ctype-mb.c:
  Adding explicit cast to return type
  in pointer substructions to avoid
  warnings from some compilers.
strings/ctype-simple.c:
  Adding explicit cast to return type
  in pointer substructions to avoid
  warnings from some compilers.
strings/ctype-sjis.c:
  Adding explicit cast to return type
  in pointer substructions to avoid
  warnings from some compilers.
strings/ctype-ucs2.c:
  Adding explicit cast to return type
  in pointer substructions to avoid
  warnings from some compilers.
strings/ctype-ujis.c:
  Adding explicit cast to return type
  in pointer substructions to avoid
  warnings from some compilers.
strings/ctype-utf8.c:
  Adding explicit cast to return type
  in pointer substructions to avoid
  warnings from some compilers.

+++ strings/ctype-big5.c
@@ -6309,7 +6309,7 @@ uint my_well_formed_len_big5(CHARSET_INFO *cs __attribute__((unused)),
       break;
     }
   }
-  return b - b0;
+  return (uint) (b - b0);
 }
